yargs-parser to normalize parsing options from repl input
