+++ README.md
@@ -11,6 +11,7 @@ A modern TypeScript API and CLI for the [Akamai NetStorage REST interface](https
 - Built-in retries, rate limiting, and dry-run mode
 - Tools for size aggregation and file validation
 - Rich CLI with matching feature parity
+- Interactive REPL with tab-completion, remote path navigation, and built-in CLI command support
 
 ## Installation
 
@@ -19,32 +20,6 @@ npm install netstorage
 ```
 
 ## Quick Start
-
-### Using the API
-
-```ts
-import { createConfig, uploadDirectory, tree } from 'netstorage';
-
-const config = createConfig({
-  key: process.env.NETSTORAGE_API_KEY!,
-  keyName: process.env.NETSTORAGE_API_KEYNAME!,
-  host: process.env.NETSTORAGE_HOST!,
-  cpCode: process.env.NETSTORAGE_CP_CODE!,
-});
-
-const remotePath = '/batch-media';
-
-await uploadDirectory(config, {
-  localPath: './media',
-  remotePath,
-});
-
-await tree(config, {
-  path: remotePath,
-  showSize: true,
-  showMtime: true,
-});
-```
 
 ### Using the CLI
 
@@ -80,10 +55,60 @@ npx netstorage rm -r old-uploads
 npx netstorage tree -s -M release-history
 ```
 
+### Using the REPL
+
+Launch an interactive NetStorage shell with:
+
+```bash
+npx netstorage
+```
+
+You get:
+
+- Remote directory navigation (`cd`, `ls`, `ll`, `pwd`)
+- Tab-completion for local and remote paths
+- Support for any standard CLI command like `upload`, `download`, `stat`, etc.
+- Remote context awareness (auto-resolves working directory when omitting arguments)
+
+Example session:
+
+```bash
+nst:/media> ls
+nst:/media> stat file.jpg
+nst:/media> put ./photo.jpg
+nst:/media> get remote-folder -v
+```
+
+### Using the API
+
+```ts
+import { createConfig, uploadDirectory, tree } from 'netstorage';
+
+const config = createConfig({
+  key: process.env.NETSTORAGE_API_KEY!,
+  keyName: process.env.NETSTORAGE_API_KEYNAME!,
+  host: process.env.NETSTORAGE_HOST!,
+  cpCode: process.env.NETSTORAGE_CP_CODE!,
+});
+
+const remotePath = '/batch-media';
+
+await uploadDirectory(config, {
+  localPath: './media',
+  remotePath,
+});
+
+await tree(config, {
+  path: remotePath,
+  showSize: true,
+  showMtime: true,
+});
+```
+
 ## Documentation
 
-- [API Reference](https://github.com/HeavyMedl/netstorage/blob/main/docs/API.md) — TypeScript interface for interacting with NetStorage programmatically
 - [CLI Reference](https://github.com/HeavyMedl/netstorage/blob/main/docs/CLI.md) — Command-line interface for NetStorage scripting
+- [API Reference](https://github.com/HeavyMedl/netstorage/blob/main/docs/API.md) — TypeScript interface for interacting with NetStorage programmatically
 
 ## Disclaimer
